+++ lib/inc/twins_string.hpp
@@ -41,6 +41,8 @@ public:
     void appendVFmt(const char *fmt, va_list ap);
     /** @brief Trim string that is too long to fit; optionally append ellipsis ... at the \p trimPos */
     String& trim(int16_t trimPos, bool addEllipsis = false, bool ignoreESC = false);
+    /** @brief Trim the string bufer to be no longer than \p trimLen bytes */
+    String& trimBuff(int16_t trimLen);
     /** @brief Strip white characters from left/right */
     String& strip(bool left = true, bool right = true);
     /** @brief Erase \p len characters from string at \p pos */